pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.8"
description = ""
authors = ["Ivan Vekhov <ivekhov@gmail.com>"]
readme = "README.md"
packages = [
    {include = "page_analyzer"},
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^22.0.0"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
validators = "^0.34.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
